requirements.txt
# ICICLE-Benchmark V2 Requirements

# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Data handling
pandas>=1.3.0
pillow>=8.3.0

# Configuration and logging
pyyaml>=5.4.0
tqdm>=4.62.0

# Visualization and analysis
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional dependencies for specific features
# Uncomment if using specific models or features
# open_clip_torch>=2.20.0  # For CLIP models
# transformers>=4.20.0     # For transformer models
# timm>=0.6.0              # For additional vision models

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
